Use BlankLayout component for restrict render VerticalLayout and HorizontalLayout when inside Login and Register (video 10)
