+++ src/pages/login.tsx
@@ -2,6 +2,8 @@
 
 import { NextPage } from 'next';
 import LoginPage from 'src/views/pages/login';
+import BlankLayout from 'src/views/layouts/BlankLayout';
+import { ReactNode } from 'react';
 
 type TProps = {};
 
@@ -10,3 +12,11 @@ const Login: NextPage<TProps> = () => {
 };
 
 export default Login;
+
+/**
+ * Code bên dưới dùng để render ra layout rỗng khi load component Login
+ * Layout rỗng chứa component Login
+ * Ngăn chặn load VerticalLayout (sidebar) và HorizontalLayout (header)
+ * getLayout là biến nằm ở file src/pages/_app.tsx
+ */
+Login.getLayout = (page: ReactNode) => <BlankLayout>{page}</BlankLayout>;
