+++ src/views/layouts/UserLayout.tsx
@@ -1,60 +1,18 @@
 import * as React from 'react';
-import { styled, createTheme, ThemeProvider } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
-import MuiDrawer from '@mui/material/Drawer';
 import Box from '@mui/material/Box';
-import MuiAppBar, { AppBarProps as MuiAppBarProps } from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
-import List from '@mui/material/List';
-import Typography from '@mui/material/Typography';
-import Divider from '@mui/material/Divider';
-import IconButton from '@mui/material/IconButton';
-import Badge from '@mui/material/Badge';
 import Container from '@mui/material/Container';
-import Grid from '@mui/material/Grid';
-import Paper from '@mui/material/Paper';
-import Link from '@mui/material/Link';
-import { mainListItems, secondaryListItems } from './listItems';
 import { NextPage } from 'next';
+import HorizontalLayout from 'src/views/layouts/HorizontalLayout';
 import VerticalLayout from 'src/views/layouts/VerticalLayout';
 
 const drawerWidth: number = 240;
-
-interface AppBarProps extends MuiAppBarProps {
-    open?: boolean;
-}
 
 type TProps = {
     children: React.ReactNode;
 };
 
-const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' })(
-    ({ theme, open }) => ({
-        '& .MuiDrawer-paper': {
-            position: 'relative',
-            whiteSpace: 'nowrap',
-            width: drawerWidth,
-            transition: theme.transitions.create('width', {
-                easing: theme.transitions.easing.sharp,
-                duration: theme.transitions.duration.enteringScreen,
-            }),
-            boxSizing: 'border-box',
-            ...(!open && {
-                overflowX: 'hidden',
-                transition: theme.transitions.create('width', {
-                    easing: theme.transitions.easing.sharp,
-                    duration: theme.transitions.duration.leavingScreen,
-                }),
-                width: theme.spacing(7),
-                [theme.breakpoints.up('sm')]: {
-                    width: theme.spacing(9),
-                },
-            }),
-        },
-    })
-);
-
-// 21p30s
 const UserLayout: NextPage<TProps> = ({ children }) => {
     const [open, setOpen] = React.useState(true);
     const toggleDrawer = () => {
@@ -70,27 +28,11 @@ const UserLayout: NextPage<TProps> = ({ children }) => {
                 toggleDrawer={toggleDrawer}
             />
 
-            <Drawer
-                variant="permanent"
+            <HorizontalLayout
                 open={open}
-            >
-                <Toolbar
-                    sx={{
-                        display: 'flex',
-                        alignItems: 'center',
-                        justifyContent: 'flex-end',
-                        px: [1],
-                    }}
-                >
-                    <IconButton onClick={toggleDrawer}>{/*<ChevronLeftIcon />*/}</IconButton>
-                </Toolbar>
-                <Divider />
-                <List component="nav">
-                    {mainListItems}
-                    <Divider sx={{ my: 1 }} />
-                    {secondaryListItems}
-                </List>
-            </Drawer>
+                toggleDrawer={toggleDrawer}
+            />
+
             <Box
                 component="main"
                 sx={{
