+++ src/views/layouts/VerticalLayout.tsx
@@ -1,10 +1,12 @@
 import * as React from 'react';
 import { styled } from '@mui/material/styles';
-import MuiAppBar, { AppBarProps as MuiAppBarProps } from '@mui/material/AppBar';
+import MuiDrawer from '@mui/material/Drawer';
+import { AppBarProps as MuiAppBarProps } from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
-import Typography from '@mui/material/Typography';
+import List from '@mui/material/List';
+import Divider from '@mui/material/Divider';
 import IconButton from '@mui/material/IconButton';
-import Badge from '@mui/material/Badge';
+import { mainListItems, secondaryListItems } from './listItems';
 import { NextPage } from 'next';
 
 const drawerWidth: number = 240;
@@ -18,66 +20,55 @@ type TProps = {
     toggleDrawer: () => void;
 };
 
-const AppBar = styled(MuiAppBar, {
-    shouldForwardProp: (prop) => prop !== 'open',
-})<AppBarProps>(({ theme, open }) => ({
-    zIndex: theme.zIndex.drawer + 1,
-    transition: theme.transitions.create(['width', 'margin'], {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.leavingScreen,
-    }),
-    ...(open && {
-        marginLeft: drawerWidth,
-        width: `calc(100% - ${drawerWidth}px)`,
-        transition: theme.transitions.create(['width', 'margin'], {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.enteringScreen,
-        }),
-    }),
-}));
+const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' })(
+    ({ theme, open }) => ({
+        '& .MuiDrawer-paper': {
+            position: 'relative',
+            whiteSpace: 'nowrap',
+            width: drawerWidth,
+            transition: theme.transitions.create('width', {
+                easing: theme.transitions.easing.sharp,
+                duration: theme.transitions.duration.enteringScreen,
+            }),
+            boxSizing: 'border-box',
+            ...(!open && {
+                overflowX: 'hidden',
+                transition: theme.transitions.create('width', {
+                    easing: theme.transitions.easing.sharp,
+                    duration: theme.transitions.duration.leavingScreen,
+                }),
+                width: theme.spacing(7),
+                [theme.breakpoints.up('sm')]: {
+                    width: theme.spacing(9),
+                },
+            }),
+        },
+    })
+);
 
 const VerticalLayout: NextPage<TProps> = ({ open, toggleDrawer }) => {
     return (
-        <AppBar
-            position="absolute"
+        <Drawer
+            variant="permanent"
             open={open}
         >
             <Toolbar
                 sx={{
-                    pr: '24px', // keep right padding when drawer closed
+                    display: 'flex',
+                    alignItems: 'center',
+                    justifyContent: 'flex-end',
+                    px: [1],
                 }}
             >
-                <IconButton
-                    edge="start"
-                    color="inherit"
-                    aria-label="open drawer"
-                    onClick={toggleDrawer}
-                    sx={{
-                        marginRight: '36px',
-                        ...(open && { display: 'none' }),
-                    }}
-                >
-                    {/*<MenuIcon />*/}
-                </IconButton>
-                <Typography
-                    component="h1"
-                    variant="h6"
-                    color="inherit"
-                    noWrap
-                    sx={{ flexGrow: 1 }}
-                >
-                    Dashboard
-                </Typography>
-                <IconButton color="inherit">
-                    <Badge
-                        badgeContent={4}
-                        color="secondary"
-                    >
-                        {/*<NotificationsIcon />*/}
-                    </Badge>
-                </IconButton>
+                <IconButton onClick={toggleDrawer}>{/*<ChevronLeftIcon />*/}</IconButton>
             </Toolbar>
-        </AppBar>
+            <Divider />
+            <List component="nav">
+                {mainListItems}
+                <Divider sx={{ my: 1 }} />
+                {secondaryListItems}
+            </List>
+        </Drawer>
     );
 };
